add Ai-chat

+++ member/src/main/java/com/ml/member/service/impl/MemberServiceImpl.java
@@ -92,7 +92,7 @@ public class MemberServiceImpl extends ServiceImpl<MemberDao, MemberEntity> impl
     @Override
     public void checkPhoneUnique(String phone) throws PhoneException {
 
-        Integer phoneCount = this.baseMapper.selectCount(new QueryWrapper<MemberEntity>().eq("mobile", phone));
+        int phoneCount = Math.toIntExact(this.baseMapper.selectCount(new QueryWrapper<MemberEntity>().eq("mobile", phone)));
 
         if (phoneCount > 0) {
             throw new PhoneException();
@@ -103,7 +103,7 @@ public class MemberServiceImpl extends ServiceImpl<MemberDao, MemberEntity> impl
     @Override
     public void checkUserNameUnique(String userName) throws UsernameException {
 
-        Integer usernameCount = this.baseMapper.selectCount(new QueryWrapper<MemberEntity>().eq("username", userName));
+        int usernameCount = Math.toIntExact(this.baseMapper.selectCount(new QueryWrapper<MemberEntity>().eq("username", userName)));
 
         if (usernameCount > 0) {
             throw new UsernameException();
